+++ src/main/scala/dx/cwl/Hint.scala
@@ -10,6 +10,7 @@ import org.w3id.cwl.cwl1_2.{
   InitialWorkDirRequirementImpl,
   InlineJavascriptRequirementImpl,
   InplaceUpdateRequirementImpl,
+  LoadListingEnum,
   LoadListingRequirementImpl,
   MultipleInputFeatureRequirementImpl,
   NetworkAccessImpl,
@@ -237,8 +238,9 @@ object LoadListingRequirement extends HintSchema {
   }
 
   override def apply(hint: Map[String, Any], schemaDefs: Map[String, CwlSchema]): Hint = {
-    LoadListingRequirement(value =
-      hint.get("loadListing").map(name => LoadListing.withName(name.toString))
+    LoadListingRequirement(value = hint
+      .get("loadListing")
+      .map(name => LoadListing.from(LoadListingEnum.fromDocumentVal(name.toString)))
     )
   }
 }
@@ -337,16 +339,14 @@ object SoftwareRequirement extends HintSchema {
 
 sealed trait InitialWorkDirEntry
 case class ValueInitialWorkDirEntry(value: CwlValue) extends InitialWorkDirEntry
-case class DirInitialWorkDirEntry(entry: CwlValue,
-                                  entryName: Option[CwlValue],
-                                  writable: Option[Boolean])
+case class DirInitialWorkDirEntry(entry: CwlValue, entryName: Option[CwlValue], writable: Boolean)
     extends InitialWorkDirEntry
 
 object DirInitialWorkDirEntry {
   def apply(dirent: DirentImpl, schemaDefs: Map[String, CwlSchema]): DirInitialWorkDirEntry = {
     val entry = CwlValue(dirent.getEntry, schemaDefs)
     val entryName = translateOptionalObject(dirent.getEntryname).map(CwlValue(_, schemaDefs))
-    val writable = translateOptional[java.lang.Boolean](dirent.getWritable).map(_.booleanValue())
+    val writable = translateOptional[java.lang.Boolean](dirent.getWritable).exists(_.booleanValue())
     DirInitialWorkDirEntry(entry, entryName, writable)
   }
 
@@ -359,10 +359,11 @@ object DirInitialWorkDirEntry {
             schemaDefs
         ),
         entryName = dirent.get("entryname").map(CwlValue(_, schemaDefs)),
-        writable = dirent.get("writable").map {
-          case s: String  => s.toBoolean
-          case b: Boolean => b
-          case other      => throw new Exception(s"invalid writable value ${other}")
+        writable = dirent.get("writable") match {
+          case Some(s: String)  => s.toBoolean
+          case Some(b: Boolean) => b
+          case None             => false
+          case other            => throw new Exception(s"invalid writable value ${other}")
         }
     )
   }