+++ src/main/scala/dx/cwl/Parser.scala
@@ -64,7 +64,7 @@ case class Parser(baseUri: Option[URI] = None,
         } else if (fields.contains("$graph")) {
           val JsArray(graph) = fields("$graph")
           graph.collectFirst {
-            case JsObject(fields) if fields.get("id").contains(JsString("#main")) =>
+            case JsObject(fields) if fields.get("id").contains(JsString(Identifier.MainFrag)) =>
               val JsString(cls) = fields("class")
               (version, cls)
           }
@@ -152,9 +152,9 @@ case class Parser(baseUri: Option[URI] = None,
             case ((primaryProc, accu), (id, _)) if accu.contains(id) => (primaryProc, accu)
             case ((primaryProc, accu), (id, rawProc)) =>
               (id, primaryProc) match {
-                case (id, Some(_)) if id.frag.contains("main") =>
+                case (id, Some(_)) if id.frag.contains(Identifier.Main) =>
                   throw new Exception("more than one process has ID frag='main'")
-                case (id, None) if id.frag.contains("main") =>
+                case (id, None) if id.frag.contains(Identifier.Main) =>
                   val (proc, newAccu) =
                     parse(rawProc,
                           source,
@@ -185,18 +185,25 @@ case class Parser(baseUri: Option[URI] = None,
     * Parses a CWL document from a file.
     * @note currently, only CommandLineTool documents are supported.
     * @param path path to the document
-    * @param defaultFrag tool/workflow name, in case it is not specified in the document
-    *             if not specified, the name of the file without .cwl is used
+    * @param defaultFrag tool/workflow name, in case it is not specified in the document.
+    *                    If not specified, the name of the file without .cwl is used.
+    * @param isPacked whether the input is in packed form; ignored if the input is a JSON
+    *                 file with a top-level "\$graph" element.
     * @return a [[Process]]
     */
-  def parseFile(path: Path, defaultFrag: Option[String] = None): (Process, Document) = {
+  def parseFile(path: Path,
+                defaultFrag: Option[String] = None,
+                isPacked: Boolean = false): (Process, Document) = {
     if (cache.contains(path)) {
       cache(path)
     } else {
-      val result = parse(RootLoader.loadDocument(path, normalizedBaseUri, loadingOptions.orNull),
-                         Some(path),
-                         defaultNamespace = Some(normalizedBaseUri),
-                         defaultFrag)
+      val result = parse(
+          RootLoader.loadDocument(path, normalizedBaseUri, loadingOptions.orNull),
+          Some(path),
+          defaultNamespace = Some(normalizedBaseUri),
+          defaultFrag,
+          isGraph = isPacked
+      )
       cache = cache + (path -> result)
       result
     }
@@ -207,13 +214,20 @@ case class Parser(baseUri: Option[URI] = None,
     * @note currently, only CommandLineTool documents are supported.
     * @param sourceCode path to the document
     * @param defaultFrag tool/workflow name, in case it is not specified in the document
+    * @param isPacked whether the input is in packed form; ignored if the input is a JSON
+    *                 file with a top-level "\$graph" element.
     * @return a [[Process]]
     */
-  def parseString(sourceCode: String, defaultFrag: Option[String] = None): (Process, Document) = {
-    parse(RootLoader.loadDocument(sourceCode, normalizedBaseUri, loadingOptions.orNull),
-          None,
-          defaultNamespace = Some(normalizedBaseUri),
-          defaultFrag)
+  def parseString(sourceCode: String,
+                  defaultFrag: Option[String] = None,
+                  isPacked: Boolean = false): (Process, Document) = {
+    parse(
+        RootLoader.loadDocument(sourceCode, normalizedBaseUri, loadingOptions.orNull),
+        None,
+        defaultNamespace = Some(normalizedBaseUri),
+        defaultFrag,
+        isGraph = isPacked
+    )
   }
 
   def parseImport(relPath: String, dependencies: Document = Document.empty): (Process, Document) = {